Fixed save error
There was a problem with the filename generation. generateUniqueFilename() inserts the date in the filename with the format dd/mm/yyyy. Of course, forward slashes (/) are not valid for filenames. To fix this, this file changes line 102:
return String.format("%s_%s.txt", title.replaceAll("[^a-zA-Z0-9\\s]", ""), dateString + "_" + timeString); // Sanitize title and combine
and adds a replace() to dateString.
return String.format("%s_%s.txt", title.replaceAll("[^a-zA-Z0-9\\s]", ""), dateString.replace("/", "-") + "_" + timeString); // Sanitize title and combine