Added ability to view saved journals
I used ChatGPT to add a button to view saved journals. It searches the directory for any file ending with .txt and displays the results. It currently displays all saved journals, regardless of date saved. The visual design also needs to be fixed as it clashes with the rest of the app.